fix quick actions triggering on different tabs
