+++ apps/expo/package.json
@@ -88,7 +88,7 @@
     "react-native-mmkv": "^2.11.0",
     "react-native-pager-view": "6.2.2",
     "react-native-purchases": "^7.4.0",
-    "react-native-reanimated": "~3.5.4",
+    "react-native-reanimated": "~3.6.0",
     "react-native-safe-area-context": "4.7.4",
     "react-native-screens": "~3.27.0",
     "react-native-svg": "14.0.0",
@@ -112,7 +112,7 @@
     "@graysky/tsconfig": "0.1.0",
     "@types/babel__core": "^7.20.5",
     "@types/mime": "^3.0.4",
-    "@types/react": "^18.2.38",
+    "@types/react": "^18.2.39",
     "@types/webpack-env": "^1.18.4",
     "eslint": "^8.54.0",
     "postcss": "^8.4.31",
