+++ apps/nextjs/package.json
@@ -33,7 +33,7 @@
     "@graysky/prettier-config": "0.1.0",
     "@graysky/tsconfig": "0.1.0",
     "@types/node": "^20.10.0",
-    "@types/react": "^18.2.38",
+    "@types/react": "^18.2.39",
     "@types/react-dom": "^18.2.17",
     "autoprefixer": "^10.4.16",
     "dotenv-cli": "^7.3.0",
